Multivariate Data Handling with S4 Classes and Sparse Matrices

An S4 update of the 'mefa' package using sparse matrices for enhanced efficiency.

Details

An S4 update of the 'mefa' package using sparse matrices for enhanced efficiency. Sparse array-like objects are supported via lists of sparse matrices.

Main functions: Xtab, Mefa.

Accessor and replacement functions: xtab, samp, taxa.

Methods: mbind, groupSums, groupMeans.

Coercion methods and virtual classes defined for cross compatibility with the mefa package. S4 object classes are described in Mefa help page.

The vignette vignette("mefa4") gives an overview of the package, gives a comparison of S3 and S4 object classes, and presents a performance review.

References

Solymos P. (2008) mefa: an R package for handling and reporting count data. Community Ecology 9, 125--127.

Solymos P. (2009) Processing ecological data in R with the mefa package. Journal of Statistical Software 29(8), 1--28. http://www.jstatsoft.org/v29/i08/

http://mefa.r-forge.r-project.org/

See also

Examples

if (FALSE) {
vignette("mefa4")
}